refactor: reorder cursor location within package
